// types.ts
export enum Role {
  ADMIN = 'Administrador',
  COORDINATOR = 'Coordinador académico',
  TEACHER = 'Profesor',
  STUDENT = 'Alumno',
  FINANCIAL_MANAGER = 'Gestor Financiero',
}
// ...
export interface SubMenuItem {
  title: string;
  path: string;
  icon?: React.ReactNode;
  roles?: Role[]; // Roles que pueden ver este sub-item
}

/**
 * Representa un elemento principal de navegación de la barra lateral.
 */
export interface SideNavItem {
  title: string;
  path: string;
  icon: React.ReactNode;
  roles: Role[]; // Qué roles pueden ver este elemento
  submenu?: SubMenuItem[];
}